requirements/launchpad.txt
# Please keep requirements in this file sorted alphabetically and
# case-insensitively.  Local forks should use +lpN suffixes and be annotated
# with information about where to find the code for the fork.
#
# Don't list entries from ztk-versions.cfg here unless overriding their
# versions; they will be included automatically.

ampoule==24.10.0
amqp==2.6.1
annotated-types==0.7.0
anyjson==0.3.3
appdirs==1.4.3
asn1crypto==0.23.0
attrs==19.3.0
Automat==20.2.0
backcall==0.2.0
backports.functools-lru-cache==1.5
# ztk-versions.cfg uses 3.2.0 on Python 3, but that drops support for Python
# 3.5.  Pin to 3.1.7 until we no longer care about xenial.
bcrypt==3.1.7
beautifulsoup4==4.12.3
billiard==3.6.4.0
bleach==6.1.0
bleach-allowlist==1.0.3
breezy==3.2.0
brz-builder==0.7.4
bson==0.5.9
boto3==1.35.71
botocore==1.35.71
celery==4.4.7
Chameleon==3.6.2
configobj==5.0.6
contextvars==2.4
constantly==15.1.0
craft-platforms==0.8.0
cryptography==2.7
cssselect==0.9.1
cssutils==2.11.1
Cython==0.29.17
d2to1==0.2.12
debtcollector==1.22.0
decorator==5.1.1
defusedxml==0.8.0rc2
difftacular==0.1
distro==1.9.0
dkimpy==1.1.8
dnspython==2.6.1
# https://code.launchpad.net/~launchpad/dohq-artifactory/+git/dohq-artifactory/+ref/0.7.630
# lp1 Fix the version determination code in setup.py.
dohq-artifactory==0.7.630+lp1
dulwich==0.19.16
eggtestinfo==0.3
enum34==1.1.6
fastimport==0.9.8
feedparser==6.0.11
fixtures==4.2.4.post1
flit-core==3.3.0
FormEncode==1.3.1
geoip2==2.9.0
grokcore.component==3.1
gunicorn==20.1.0
hatch-vcs==0.4.0
httplib2==0.8
hyperlink==18.0.0
immutables==0.14
importlib==1.0.2
importlib-metadata==8.5.0
incremental==21.3.0
ipython==8.12.3
ipython-genutils==0.2.0
iso8601==0.1.12
jedi==0.17.2
jmespath==0.10.0
jsautobuild==0.2
keyring==0.6.2
keystoneauth1==4.1.0
kombu==4.6.11
launchpad-buildd==206
launchpadlib==2.1.0
lazr.batchnavigator==1.3.1
lazr.config==2.2.3
lazr.delegates==2.0.4
lazr.enum==1.2.1
lazr.jobrunner==0.17
lazr.lifecycle==1.2.1
lazr.restful==2.0.2
lazr.restfulclient==0.14.5
lazr.sshserver==0.1.13
lazr.uri==1.0.6
lxml==5.3.0
loggerhead==2.0.1
lpjsmin==0.7
m2r==0.1.13
Markdown==3.7
martian==1.3.post1
maxminddb==1.5.1
mistune==0.8.3
monotonic==1.5
more-itertools==10.5.0
msgpack==1.0.2
multipart==1.3.1
netaddr==0.7.19
netifaces==0.11.0
oauth==1.0
oauthlib==3.1.0
oops==0.0.14
oops-amqp==0.2.0
oops-datedir-repo==0.0.24
oops-datedir2amqp==0.1.0
oops-timeline==0.0.3
oops-twisted==0.0.8
oops-wsgi==0.0.15
os-service-types==1.7.0
oslo.config==7.0.0
oslo.i18n==3.25.1
oslo.serialization==2.29.3
oslo.utils==3.42.1
packaging==24.2
paramiko==2.4.1
parso==0.7.1
Paste==3.5.2
PasteDeploy==2.1.0
pathlib2==2.3.2
patiencediff==0.2.2
pexpect==4.8.0
pgbouncer==0.0.9
pickleshare==0.7.5
pkginfo==1.11.2
prettytable==0.7.2
psutil==7.0.0
psycopg2==2.8.6
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pygettextpo==0.3
# lp:~launchpad/pygpgme/pthread
# lp1 Link against libgpgme-pthread for thread-safety.
pygpgme==0.3+lp1; python_version < '3.9'
pygpgme==0.6; python_version >= '3.9'
PyHamcrest==1.9.0
pyinotify==0.9.4
PyJWT==1.7.1
pymacaroons==0.13.0
pymemcache==4.0.0
PyNaCl==1.3.0
pyOpenSSL==17.5.0
pyparsing==3.1.4
pystache==0.6.6
python-dateutil==2.9.0.post0
python-debian==0.1.49
python-keystoneclient==3.21.0
python3-openid==3.2
# ztk-versions.cfg uses 1.4.0 on Python 3, but that does not support 3.12.
# Pin to 1.4.4 to override it.
python-subunit==1.4.4
python-swiftclient==4.6.0
PyYAML==6.0.2
rabbitfixture==0.5.3
requests==2.32.3
requests-file==1.4.3
requests-toolbelt==0.9.1
responses==0.17.0
rfc3986==1.5.0
s3transfer==0.10.4
secure-cookie==0.1.0
selenium==3.141.0
service-identity==18.1.0
setproctitle==1.3.4
setuptools-git==1.2
setuptools-scm==3.4.3
sgmllib3k==1.0.0
soupmatchers==0.4
soupsieve==2.6
statsd==3.3.0
stevedore==1.32.0
# lp:~launchpad-committers/storm/lp
storm==0.26+lp418
subprocess32==3.2.6
tenacity==9.0.0
testresources==0.2.7
testscenarios==0.4
testtools==2.7.2
timeline==0.0.7
toml==0.10.2
transaction==3.0.1
treq==18.6.0
# lp:~launchpad/twisted:lp-backport
Twisted==20.3.0+lp10
# Python 3.8 and twisted 20.3.0+lp9
txfixtures==0.4.3
# Python 3.10+ and twisted 24.11
# XXX jugmac00 2025-02-18: the twisted update still needs to be done
txfixtures==0.5.2; python_version >= "3.10"
txpkgupload==0.5
typing_extensions==4.12.2; python_version >= "3.9"
urllib3==1.26.20
van.testing==3.0.0
vine==1.3.0
virtualenv-tools3==3.1.1
wadllib==1.3.6
waitress==2.1.2
wcwidth==0.2.6
webencodings==0.5.1
WebOb==1.8.9
WebTest==2.0.35
Werkzeug==1.0.1
wrapt==1.12.1
wsgi-intercept==1.9.2
WSGIProxy2==0.4.6
z3c.pt==3.2.0
z3c.ptcompat==2.2.0
zipp==3.20.2
zope.app.http==4.0.1
zope.app.publication==4.3.1
zope.app.publisher==4.2.0
zope.publisher==6.0.2+lp1
zope.testbrowser==5.5.1
# lp:~launchpad-committers/zope.testrunner:launchpad-6.6.1+lp1
zope.testrunner==6.6.1+lp1
zope.vocabularyregistry==1.1.1
zstandard==0.15.2
